+++ test/dummy/rails52_sprockets4/app/controllers/admin_controller.rb
@@ -1,2 +1,4 @@
 class AdminController < Super::ApplicationController
+  class AdminControls < Super::Controls
+  end
 end