+++ test/dummy/rails60_webpacker4/app/controllers/admin/members_controller.rb
@@ -11,7 +11,7 @@ module Admin
       Controls.new
     end
 
-    class Controls < Super::Controls
+    class Controls < AdminControls
       def title
         Member.name.pluralize
       end