+++ test/dummy/rails61_webpacker5/app/controllers/admin/ships_controller.rb
@@ -6,7 +6,7 @@ module Admin
       Controls.new
     end
 
-    class Controls < Super::Controls
+    class Controls < AdminControls
       def title
         Ship.name.pluralize
       end